+++ reactproyecto/src/App.js
@@ -1,5 +1,5 @@
 import React from "react"
-import {Header} from "./componentes/Header"
+import { Header } from "./componentes/Header/index.js"
 import 'boxicons'
 import logo from "../src/images/Logo.jpg"
 import Carousel from "./componentes/carousel"
@@ -7,34 +7,19 @@ import Carousel from "./componentes/carousel"
 function App() {
   return (
     <div className="App">
-
       <div className="row bg-primary">
-       
-          <Header className="App-header">
-            <img src={logo} className="App-logo" alt="logo" />
-            <p>
-              Edit <code>src/App.js</code> and save to reload.
-            </p>
-            <a
-              className="App-link"
-              href="https://reactjs.org"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-            </a>
-          </Header>
-
+        <Header className="header"></Header>
       </div>
 
       <div className="row">
-        <div className="col">
-          <Carousel/>
+        <div className="container-fluid">
+          <Carousel />
         </div>
       </div>
 
     </div>
-    
-    
+
+
   );
 }
