+++ reactproyecto/src/api/db.json
@@ -29,6 +29,36 @@
             "src":"img5.jpg",
             "title":"imagen-5",
             "caption":""
+        },
+        {
+            "id":6,
+            "src":"img1.jpg",
+            "title":"imagen-1",
+            "caption":""
+        },
+        {
+            "id":7,
+            "src":"img2.jpg",
+            "title":"imagen-2",
+            "caption":""
+        },
+        {
+            "id":8,
+            "src":"img3.jpg",
+            "title":"imagen-3",
+            "caption":""
+        },
+        {
+            "id":9,
+            "src":"img4.jpg",
+            "title":"imagen-4",
+            "caption":""
+        },
+        {
+            "id":10,
+            "src":"img5.jpg",
+            "title":"imagen-10",
+            "caption":""
         }
     ]
 }
