+++ reactproyecto/src/css/estilos.css
@@ -1,0 +1,5 @@
+
+/* carousel */
+.item img{
+    pointer-events: none;
+}
