+++ reactproyecto/src/index.css
@@ -33,7 +33,7 @@ header{
   justify-content: space-around;
   align-items: center;
   flex-wrap: wrap;
-  position: fixed;
+  /* position: fixed; */
   box-sizing: border-box;
   background-color: white;
   box-shadow: 0 5px 10px #b8cac1c5;
@@ -179,11 +179,4 @@ header .cart span{
   header ul{
     display: none;
   }
-}
-
-
-/*----------------------------------CAROUSEL--------------------------------*/
-.miCarousel{
-  width: 150px;
-  height: auto;
 }
